- Adaptor to implement the features
