+++ ruleunit-reactive-quarkus/README.md
@@ -31,18 +31,4 @@ You will see the result in "alerts" topic via Kafdrop (http://localhost:9000).
 ```
 
 
-## Test with REST
-
-### POST /warnings
-
-Returns warning level alerts caused by incoming events:
-
-```sh
-curl -X POST -H 'Accept: application/json' -H 'Content-Type: application/json' -d '{"eventData":[{"type":"temperature","value":25}, {"type":"temperature","value":35}]}' http://localhost:8080/warnings
-```
-
-Example response:
-
-```json
-[{"severity":"warning","message":"Event [type=temperature, value=40]"}]
-```
+Note that this example doesn't expose REST endpoints.
