--- ruleunit-reactive-quarkus/src/main/java/org/kie/kogito/alerts/AlertingServiceEmitter.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kie.kogito.alerts;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-
-import org.eclipse.microprofile.reactive.messaging.Channel;
-import org.eclipse.microprofile.reactive.messaging.Emitter;
-
-@ApplicationScoped
-public class AlertingServiceEmitter {
-
-    @Inject
-    @Channel("alerts")
-    Emitter<Alert> emitter;
-
-    public void send(Alert alert) {
-        System.out.println("AlertingServiceEmitter : send " + alert);
-        emitter.send(alert);
-    }
-}
